Standard format corrections for controllers
